+++ integration_tests/suite/base/test_queues.py
@@ -287,6 +287,10 @@ def test_get(queue):
 
         extensions=empty(),
         schedules=empty(),
+        members=has_entries(
+            agents=empty(),
+            users=empty(),
+        ),
     ))