+++ xivo_confd/plugins/queue_member/service.py
@@ -1,41 +1,87 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2016 Avencall
+# Copyright 2016-2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
-from xivo_dao.resources.queue_members import dao as queue_member_dao
-from xivo_confd.plugins.queue_member.notifier import build_notifier
-from xivo_confd.plugins.queue_member.validator import build_validator
+from xivo_dao.helpers import errors
+from xivo_dao.resources.agent import dao as agent_dao_module
+from xivo_dao.resources.queue import dao as queue_dao_module
+
+from .notifier import build_notifier
+from .validator import build_validator_member_user
 
 
 class QueueMemberService(object):
 
-    def __init__(self, dao, validator, notifier):
-        self.dao = dao
-        self.validator = validator
+    def __init__(self, queue_dao, agent_dao, validator_member_user, notifier):
+        self.queue_dao = queue_dao
+        self.agent_dao = agent_dao
         self.notifier = notifier
+        self.validator_member_user = validator_member_user
 
-    def get(self, queue_id, agent_id):
-        self.validator.validate_get_agent_queue_association(queue_id, agent_id)
-        return self.dao.get_by_queue_id_and_agent_id(queue_id, agent_id)
+    def get_member_agent(self, queue, agent):
+        member = self.find_member_agent(queue, agent)
+        if not member:
+            raise errors.not_found('QueueMember', agent_id=agent.id, queue_id=queue.id)
+        return member
 
-    def edit(self, queue_member):
-        self.validator.validate_edit_agent_queue_association(queue_member)
-        self.dao.edit_agent_queue_association(queue_member)
-        self.notifier.edited(queue_member)
+    def get_agent(self, agent_id):
+        agent = self.agent_dao.find(agent_id)
+        if not agent:
+            raise errors.param_not_found('agent_id', 'Agent')
+        return agent
 
-    def associate(self, queue_member):
-        self.validator.validate_associate_agent_queue(queue_member.queue_id, queue_member.agent_id)
-        qm = self.dao.associate(queue_member)
-        self.notifier.associated(queue_member)
-        return qm
+    def find_member_agent(self, queue, agent):
+        for member in queue.agent_queue_members:
+            if member.agent == agent:
+                return member
+        return None
 
-    def dissociate(self, queue_member):
-        self.validator.validate_remove_agent_from_queue(queue_member.queue_id, queue_member.agent_id)
-        self.dao.remove_agent_from_queue(queue_member.agent_id, queue_member.queue_id)
-        self.notifier.dissociated(queue_member)
+    def find_member_user(self, queue, user):
+        for member in queue.user_queue_members:
+            if member.user == user:
+                return member
+        return None
+
+    def associate_legacy(self, queue, member):
+        if member in queue.agent_queue_members:
+            raise errors.resource_associated('Agent', 'Queue', member.agent.id, queue.id)
+        self.queue_dao.associate_member_agent(queue, member)
+        self.notifier.agent_associated(queue, member)
+
+    def associate_member_agent(self, queue, member):
+        if member in queue.agent_queue_members:
+            return
+
+        self.queue_dao.associate_member_agent(queue, member)
+        self.notifier.agent_associated(queue, member)
+
+    def dissociate_member_agent(self, queue, member):
+        if member not in queue.agent_queue_members:
+            return
+
+        self.queue_dao.dissociate_member_agent(queue, member)
+        self.notifier.agent_dissociated(queue, member)
+
+    def associate_member_user(self, queue, member):
+        if member in queue.user_queue_members:
+            return
+
+        self.validator_member_user.validate_association(queue, member)
+        self.queue_dao.associate_member_user(queue, member)
+        self.notifier.user_associated(queue, member)
+
+    def dissociate_member_user(self, queue, member):
+        if member not in queue.user_queue_members:
+            return
+
+        self.queue_dao.dissociate_member_user(queue, member)
+        self.notifier.user_dissociated(queue, member)
 
 
 def build_service():
-    return QueueMemberService(queue_member_dao,
-                              build_validator(),
-                              build_notifier())
+    return QueueMemberService(
+        queue_dao_module,
+        agent_dao_module,
+        build_validator_member_user(),
+        build_notifier(),
+    )