+++ xivo_confd/plugins/queue_member/validator.py
@@ -1,56 +1,31 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2013-2016 Avencall
+# Copyright 2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
-from xivo_dao import queue_dao, agent_dao
-from xivo_dao.helpers.exception import NotFoundError
 from xivo_dao.helpers import errors
-from xivo_dao.resources.queue_members import dao as queue_members_dao
+
+from xivo_confd.helpers.validator import ValidatorAssociation, ValidationAssociation
 
 
-class QueueMemberAssociationValidator(object):
+class QueueMemberUserAssociationValidator(ValidatorAssociation):
 
-    def validate_edit_agent_queue_association(self, queue_member):
-        if not self._queue_exists(queue_member.queue_id):
-            raise errors.not_found('Queue', queue_id=queue_member.queue_id)
-        if not self._agent_exists(queue_member.agent_id):
-            raise errors.not_found('Agent', agent_id=queue_member.agent_id)
-        queue_members_dao.get_by_queue_id_and_agent_id(queue_member.queue_id, queue_member.agent_id)
+    def validate(self, queue, member):
+        self.validate_user_has_endpoint(member.user)
+        self.validate_no_users_have_same_line(queue, member.user)
 
-    def validate_get_agent_queue_association(self, queue_id, agent_id):
-        if not self._queue_exists(queue_id):
-            raise errors.not_found('Queue', queue_id=queue_id)
-        if not self._agent_exists(agent_id):
-            raise errors.not_found('Agent', agent_id=agent_id)
+    def validate_user_has_endpoint(self, user):
+        if not user.lines:
+            raise errors.missing_association('User', 'Line',
+                                             user_uuid=user.uuid)
 
-    def validate_associate_agent_queue(self, queue_id, agent_id):
-        if not self._queue_exists(queue_id):
-            raise errors.not_found('Queue', queue_id=queue_id)
-        if not self._agent_exists(agent_id):
-            raise errors.param_not_found('agent_id', 'Agent')
-        try:
-            queue_members_dao.get_by_queue_id_and_agent_id(queue_id, agent_id)
-            raise errors.resource_associated('Agent', 'Queue',
-                                             agent_id, queue_id)
-        except NotFoundError:
-            pass
-
-    def validate_remove_agent_from_queue(self, queue_id, agent_id):
-        if not self._queue_exists(queue_id):
-            raise errors.not_found('Queue', queue_id=queue_id)
-        if not self._agent_exists(agent_id):
-            raise errors.not_found('Agent', agent_id=agent_id)
-        queue_members_dao.get_by_queue_id_and_agent_id(queue_id, agent_id)
-
-    def _queue_exists(self, queue_id):
-        try:
-            return queue_dao.get(queue_id) is not None
-        except LookupError:
-            return False
-
-    def _agent_exists(self, agent_id):
-        return agent_dao.get(agent_id) is not None
+    def validate_no_users_have_same_line(self, queue, user):
+        all_lines = [member.user.lines[0] for member in queue.user_queue_members]
+        if user.lines[0] in all_lines:
+            raise errors.not_permitted('Cannot associate different users with the same line',
+                                       line_id=user.lines[0].id)
 
 
-def build_validator():
-    return QueueMemberAssociationValidator()
+def build_validator_member_user():
+    return ValidationAssociation(
+        association=[QueueMemberUserAssociationValidator()],
+    )