+++ CHANGELOG.md
@@ -4,6 +4,45 @@ Changelog
 18.08
 -----
 
+* A new API for associating user with a queue has been added:
+
+  * PUT `/1.1/queues/<queue_id>/members/users/<user_id>`
+  * DELETE `/1.1/queues/<queue_id>/members/users/<user_id>`
+
+* The following URLs have been deprecated. Please use the new API instead:
+
+  * POST `/1.1/queues/<queue_id>/members/agents`
+  * GET `/1.1/queues/<queue_id>/members/agents/<agent_id>`
+
+* New readonly parameters have been added to the queue resource:
+
+  * `members`
+
+    * `agents`
+    * `users`
+
+* New readonly parameters have been added to the agent resource:
+
+  * `queues`
+
+* New readonly parameters have been added to the user resource:
+
+  * `queues`
+
+* New parameter have been added to the queue-member agent association:
+
+  * `priority`
+
+* The following endpoint does not return an error (400) when the resources are not already associated.
+  Instead, it associate resources and return a successful association code (204).
+
+  * PUT `/1.1/queues/<queue_id>/members/agents/<agent_id>`
+
+* The following endpoint does not return a (400) error when the resources are not associated.
+  Instead, it returns a successful dissociation code (204).
+
+  * DELETE `/1.1/queues/<queue_id>/members/agents/<agent_id>`
+
 * A new API for associating a queue with a schedule has been added:
 
   * DELETE `/1.1/queues/<queue_id>/schedules/<schedule_id>`